ref impl: naming fixes and code organization
