+++ opnode/l2/input_derivation_test.go
